Binary QSAR Data

Binary QSAR Data

Value

binary_tr,binary_ukn

data frame frames with 49 columns

Details

These data are from two different sources on quantitative structure-activity relationship (QSAR) modeling and contain 49 predictors that are either 0 or 1. The training set contains 3796 samples (from the caco data set in the QSARdata package) and there are 80 unknown samples (taken from the bbb2 data in the same package).

The training set was collected from a general set used to measure permeability while the new samples are from a set of molecules used to study brain penetration (presumably for neuroscience related purposes).

Examples

data(qsar_binary)
str(binary_tr)

#> 'data.frame':	3796 obs. of  49 variables:
#>  $ pred01: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred02: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred03: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred04: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred05: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred06: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred07: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred08: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred09: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred10: int  0 1 1 1 1 1 1 1 1 1 ...
#>  $ pred11: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred12: int  0 0 1 1 1 1 1 1 1 0 ...
#>  $ pred13: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred14: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred15: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred16: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred17: int  0 0 1 1 1 1 1 1 1 0 ...
#>  $ pred18: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred19: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred20: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred21: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pred22: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred23: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred24: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pred25: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred26: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred27: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pred28: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred29: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred30: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pred31: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred32: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred33: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred34: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred35: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred36: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred37: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred38: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred39: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred40: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred41: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred42: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred43: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred44: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred45: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred46: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred47: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred48: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pred49: int  0 0 0 0 0 0 0 0 0 0 ...