OkCupid Binary Predictors

OkCupid Binary Predictors

Source

Kim (2015), "OkCupid Data for Introductory Statistics and Data Science Courses", Journal of Statistics Education, Volume 23, Number 2. http://www.amstat.org/publications/jse/contents_2015.html

Kuhn and Johnson (2020), Feature Engineering and Selection, Chapman and Hall/CRC . https://bookdown.org/max/FES/ and https://github.com/topepo/FES

Value

okc_binary_train,okc_binary_test

data frame frames with 61 columns

Details

Data originally from Kim (2015) includes a training and test set consistent with Kuhn and Johnson (2020). Predictors include ethnicity indicators and a set of keywords derived from text essay data.

Examples

data(okc_binary)
str(okc_binary_train)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	38809 obs. of  61 variables:
#>  $ software        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ engineer        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ startup         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ tech            : num  0 0 0 0 1 0 0 0 0 1 ...
#>  $ computers       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ engineering     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ computer        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ internet        : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ technology      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ science         : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ programming     : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ technical       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ web             : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ developer       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ im              : num  1 0 1 0 1 1 0 1 1 0 ...
#>  $ programmer      : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ scientist       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ code            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ stephenson      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ geek            : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ nerd            : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ lol             : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ biotech         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ matrix          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ coding          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ geeky           : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ solving         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ problems        : num  0 0 1 0 1 0 0 0 0 0 ...
#>  $ data            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ fixing          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ teacher         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ student         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ silicon         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ law             : num  0 0 0 0 0 0 0 1 0 0 ...
#>  $ mechanical      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ electronic      : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ pratchett       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ wikipedia       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ neal            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ mobile          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ math            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ lab             : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ systems         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ electronics     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ futurama        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ alot            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ solve           : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ websites        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ firefly         : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ valley          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ apps            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ lawyer          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ asian           : num  1 0 0 0 0 0 0 0 0 0 ...
#>  $ black           : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hispanic_latin  : num  0 0 0 0 0 0 0 1 0 0 ...
#>  $ indian          : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ middle_eastern  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ native_american : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ other           : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pacific_islander: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ white           : num  1 1 1 1 1 1 1 1 1 1 ...