Predict from a apd_pca

Predict from a apd_pca

# S3 method for apd_pca
score(object, new_data, type = "numeric", ...)

Arguments

object
A apd_pca object.

new_data
A data frame or matrix of new predictors.

type
A single character. The type of predictions to generate. Valid options are:

"numeric" for numeric predictions.

...
Not used, but required for extensibility.

Value

A tibble of predictions. The number of rows in the tibble is guaranteed to be the same as the number of rows in new_data.

Examples

train <- mtcars[1:20,]
test <- mtcars[21:32, -1]

# Fit
mod <- apd_pca(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
score(mod, test)

#>           PC1         PC2 hat_values     dists
#> 1  -1.1610755  0.66426939 0.10475548 1.3376660
#> 2   1.8385562  0.34509216 0.14943286 1.8706623
#> 3   1.2341928 -0.25927125 0.12332836 1.2611318
#> 4   0.4614036 -1.03206040 0.10262327 1.1305052
#> 5   1.3369513 -0.15651276 0.12715300 1.3460813
#> 6  -1.6081081  0.21723682 0.13485638 1.6227148
#> 7  -1.9844417 -0.15909677 0.16388695 1.9908090
#> 8  -1.2467753  0.57856963 0.11015729 1.3744786
#> 9  -0.1029733 -1.59643733 0.09648952 1.5997549
#> 10 -0.2314226 -0.06548219 0.06196518 0.2405085
#> 11  0.7004640 -0.79300005 0.10750881 1.0580637
#> 12 -1.6416069  0.18373806 0.13730373 1.6518574